Добавить проверку на 1 открытый ревит и отсутствие temp директории

+++ startup.py
@@ -4,20 +4,25 @@ import os
 import sys
 from pyrevit.versionmgr import updater
 import LibGit2Sharp as libgit
+from pyrevit import coreutils
+import tempfile
 
 
 def update_extensions():
-    for repo_info in updater.get_all_extension_repos():
-        try:
-            if updater.has_pending_updates(repo_info):
-                # сбрасываем репозиторий в исходное состояние
-                repo_info.repo.Reset(libgit.ResetMode.Hard, repo_info.repo.Head.Tip)
-                repo_info.repo.RemoveUntrackedFiles()
+    revit_count = coreutils.get_revit_instance_count()
+    temp_dir_exist = os.path.isdir(os.path.join(tempfile.gettempdir(), "Bim4Everyone", "dosymep"))
+    if revit_count == 1 and not temp_dir_exist:
+        for repo_info in updater.get_all_extension_repos():
+            try:
+                if updater.has_pending_updates(repo_info):
+                    # сбрасываем репозиторий в исходное состояние
+                    repo_info.repo.Reset(libgit.ResetMode.Hard, repo_info.repo.Head.Tip)
+                    repo_info.repo.RemoveUntrackedFiles()
 
-                # пытаемся обновится
-                updater.update_repo(repo_info)
-        except Exception:
-            pass
+                    # пытаемся обновится
+                    updater.update_repo(repo_info)
+            except Exception:
+                pass
 
 
 update_extensions()
